+++ central/vulnerabilityrequest/datastore/datastore.go
@@ -17,7 +17,7 @@ import (
 	pkgRocksDB "github.com/stackrox/rox/pkg/rocksdb"
 	"github.com/stackrox/rox/pkg/sac"
 	"github.com/stackrox/rox/pkg/search"
-	"github.com/stackrox/rox/pkg/testutils"
+	"github.com/stackrox/rox/pkg/testutils/utils"
 )
 
 // DataStore is an intermediary to VulnerabilityRequest storage.
@@ -70,7 +70,7 @@ func New(s store.Store, indexer index.Indexer, searcher searcher.Searcher,
 // To make this more explicit, we require passing a testing.T to this version.
 func NewForTestOnly(t *testing.T, db *pkgRocksDB.RocksDB, bleveIndex bleve.Index,
 	pendingReqCache cache.VulnReqCache, activeReqCache cache.VulnReqCache) (DataStore, error) {
-	testutils.MustBeInTest(t)
+	utils.MustBeInTest(t)
 
 	storage := rocksdb.New(db)
 	indexer := index.New(bleveIndex)
