+++ pkg/buildinfo/testutils/timestamp.go
@@ -5,12 +5,11 @@ import (
 	"time"
 
 	"github.com/stackrox/rox/pkg/buildinfo/internal/timestamp"
-	"github.com/stackrox/rox/pkg/testutils"
 )
 
 // SetBuildTimestamp sets the build timestamp in UNIX secs. This function is only intended for testing.
 func SetBuildTimestamp(t *testing.T, buildTimestamp time.Time) {
-	testutils.MustBeInTest(t)
+	utils.MustBeInTest(t)
 	timestamp.BuildTimestamp = buildTimestamp
 	timestamp.BuildTimestampParsingErr = nil
 }
