+++ pkg/testutils/utils/verify_in_test_only.go
@@ -1,4 +1,4 @@
-package testutils
+package utils
 
 import (
 	"flag"
